Revert "feat: unify segmented control sizing helpers"
