+++ OffshoreBudgeting/Views/HomeView.swift
@@ -1065,6 +1065,89 @@ private struct HomeSegmentTotalsRowView: View {
     }
 }
 
+// MARK: - Segmented control sizing helpers
+private extension View {
+    func segmentedFill() -> some View { frame(maxWidth: .infinity) }
+    func equalWidthSegments() -> some View { modifier(HomeEqualWidthSegmentsModifier()) }
+}
+
+private struct HomeEqualWidthSegmentsModifier: ViewModifier {
+    func body(content: Content) -> some View {
+#if os(iOS)
+        content.background(HomeEqualWidthSegmentApplier())
+#elseif os(macOS)
+        content.background(HomeEqualWidthSegmentApplier())
+#else
+        content
+#endif
+    }
+}
+
+#if os(iOS)
+private struct HomeEqualWidthSegmentApplier: UIViewRepresentable {
+    func makeUIView(context: Context) -> UIView {
+        let view = UIView()
+        view.isUserInteractionEnabled = false
+        DispatchQueue.main.async { applyEqualWidthIfNeeded(from: view) }
+        return view
+    }
+
+    func updateUIView(_ uiView: UIView, context: Context) {
+        DispatchQueue.main.async { applyEqualWidthIfNeeded(from: uiView) }
+    }
+
+    private func applyEqualWidthIfNeeded(from view: UIView) {
+        guard let segmented = findSegmentedControl(from: view) else { return }
+        segmented.apportionsSegmentWidthsByContent = false
+        segmented.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
+        segmented.setContentHuggingPriority(.defaultLow, for: .horizontal)
+        segmented.invalidateIntrinsicContentSize()
+    }
+
+    private func findSegmentedControl(from view: UIView) -> UISegmentedControl? {
+        var current: UIView? = view
+        while let candidate = current {
+            if let segmented = candidate as? UISegmentedControl { return segmented }
+            current = candidate.superview
+        }
+        return nil
+    }
+}
+#elseif os(macOS)
+private struct HomeEqualWidthSegmentApplier: NSViewRepresentable {
+    func makeNSView(context: Context) -> NSView {
+        let view = NSView(frame: .zero)
+        view.alphaValue = 0.0
+        DispatchQueue.main.async { applyEqualWidthIfNeeded(from: view) }
+        return view
+    }
+
+    func updateNSView(_ nsView: NSView, context: Context) {
+        DispatchQueue.main.async { applyEqualWidthIfNeeded(from: nsView) }
+    }
+
+    private func applyEqualWidthIfNeeded(from view: NSView) {
+        guard let segmented = findSegmentedControl(from: view) else { return }
+        SegmentedControlEqualWidthCoordinator.enforceEqualWidth(for: segmented)
+    }
+
+    private func findSegmentedControl(from view: NSView) -> NSSegmentedControl? {
+        // Prefer searching siblings/descendants of the immediate superview, because
+        // this representable is attached as a background.
+        guard let root = view.superview else { return nil }
+        return searchSegmented(in: root)
+    }
+
+    private func searchSegmented(in node: NSView) -> NSSegmentedControl? {
+        for sub in node.subviews {
+            if let seg = sub as? NSSegmentedControl { return seg }
+            if let found = searchSegmented(in: sub) { return found }
+        }
+        return nil
+    }
+}
+#endif
+
 // MARK: - Home Header Summary
 private struct HomeHeaderContextSummary: View {
     let summary: BudgetSummary
